refactor: replace useAdminPagination with usePaginate for consistent pagination handling
- Updated AdminApplications, AdminCompute, AdminDatabase, AdminOperations, AdminOrganizations, AdminStorage, AdminUsers components to use the new usePaginate hook for pagination instead of the deprecated useAdminPagination.
- Introduced a new usePaginate hook to manage client-side pagination logic, improving code reusability and maintainability.
- Added ApplicationSettings and People components for managing organization applications and members, respectively.
- Implemented Settings component to manage organization settings, including applications, members, and storage usage.
- Enhanced documentation and error handling in various components.
