+++ web/src/platform/routes/Admin/Organizations.tsx
@@ -15,9 +15,9 @@ import { fetchApiVoid } from '@/lib/api';
 import { useToast } from '@/hooks/use-toast';
 import { useApiQuery } from '@/hooks/use-api';
 import { useDeleteDialog } from '@/lib/utils';
+import { usePaginate } from '@/hooks/pagination';
 import { platformApiPath } from '@/lib/platform-api';
 import { organizationsQueryKey } from '@/lib/query-keys';
-import { useAdminPagination } from '@/platform/admin/pagination';
 import { DeleteConfirmation } from '@/components/dialogs/DeleteConfirmation';
 import { zOrganizationSummary } from '@/lib/generated/platform-api-v1/zod.gen';
 
@@ -40,7 +40,7 @@ export default function AdminOrganizations() {
     } = useApiQuery<OrganizationSummary[]>(platformApiPath('/organizations'), {
         parse: (value) => zOrganizationSummary.array().parse(value),
     });
-    const { pageItems, pagination } = useAdminPagination(organizations);
+    const { pageItems, pagination } = usePaginate(organizations);
     const deleteDialog = useDeleteDialog({
         title: 'Delete organization',
         mutation: deleteOrganization,
